Allow basic branch out functionality in packages
