+++ scale_build/packages/package.py
@@ -1,18 +1,16 @@
 import json
 import logging
 import os
-import shutil
 
 from scale_build.exceptions import CallError
-from scale_build.utils.git_utils import retrieve_git_remote_and_sha, retrieve_git_branch, update_git_manifest
-from scale_build.utils.logger import LoggingContext
 from scale_build.utils.run import run
-from scale_build.utils.paths import GIT_LOG_PATH, HASH_DIR, PKG_LOG_DIR, SOURCES_DIR
+from scale_build.utils.paths import HASH_DIR, PKG_LOG_DIR, SOURCES_DIR
 
 from .binary_package import BinaryPackage
 from .bootstrap import BootstrapMixin
 from .build import BuildPackageMixin
 from .clean import BuildCleanMixin
+from .git import GitPackageMixin
 from .overlay import OverlayMixin
 from .utils import (
     DEPENDS_SCRIPT_PATH, gather_build_time_dependencies, normalize_build_depends, normalize_bin_packages_depends
@@ -22,7 +20,7 @@ from .utils import (
 logger = logging.getLogger(__name__)
 
 
-class Package(BootstrapMixin, BuildPackageMixin, BuildCleanMixin, OverlayMixin):
+class Package(BootstrapMixin, BuildPackageMixin, BuildCleanMixin, GitPackageMixin, OverlayMixin):
     def __init__(
         self, name, branch, repo, prebuildcmd=None, kernel_module=False, explicit_deps=None,
         generate_version=True, predepscmd=None, deps_path=None, subdir=None, deoptions=None, jobs=None,
@@ -154,41 +152,6 @@ class Package(BootstrapMixin, BuildPackageMixin, BuildCleanMixin, OverlayMixin):
     @property
     def exists(self):
         return os.path.exists(self.source_path)
-
-    def retrieve_current_remote_origin_and_sha(self):
-        if self.exists:
-            return retrieve_git_remote_and_sha(self.source_path)
-        else:
-            return {'url': None, 'sha': None}
-
-    def update_git_manifest(self):
-        info = self.retrieve_current_remote_origin_and_sha()
-        update_git_manifest(info['url'], info['sha'])
-
-    def checkout(self, branch_override=None):
-        origin_url = self.retrieve_current_remote_origin_and_sha()['url']
-        branch = branch_override or self.branch
-        if branch == self.existing_branch and self.origin == origin_url:
-            logger.debug('Updating git repo [%s] (%s)', self.name, GIT_LOG_PATH)
-            with LoggingContext('git-checkout', 'w'):
-                run(['git', '-C', self.source_path, 'fetch', 'origin'])
-                run(['git', '-C', self.source_path, 'checkout', branch])
-                run(['git', '-C', self.source_path, 'reset', '--hard', f'origin/{branch}'])
-        else:
-            logger.debug('Checking out git repo [%s] (%s)', self.name, GIT_LOG_PATH)
-            if os.path.exists(self.source_path):
-                shutil.rmtree(self.source_path)
-            with LoggingContext('git-checkout', 'w'):
-                run(['git', 'clone', '--recurse', self.origin, self.source_path])
-                run(['git', '-C', self.source_path, 'checkout', branch])
-
-        self.update_git_manifest()
-
-    @property
-    def existing_branch(self):
-        if not self.exists:
-            return None
-        return retrieve_git_branch(self.source_path)
 
     @property
     def pkglist_hash_file_path(self):
